requirement.txt
pandas
sqlalchemy
python-dotenv
requests
b2sdk
azure-cognitiveservices-speech
google-generativeai
tenacity
tqdm
concurrent-log-handler   # optional for rotating logs
cockroachdb
psycopg2-binary
b2sdk>=2.1.0
requests>=2.28.0
tqdm>=4.64.0
retrying>=1.3.3
